decomposing stacks into constructs
