+++ lib/stacks/api/jobs.ts
@@ -15,11 +15,8 @@
 */
 
 import { App, Stack } from 'aws-cdk-lib';
-import { LayerVersion } from 'aws-cdk-lib/aws-lambda';
-import { StringParameter } from 'aws-cdk-lib/aws-ssm';
-import { MLSpacePythonLambdaFunction, registerAPIEndpoint } from '../../utils/apiFunction';
 import { ApiStackProperties } from './restApi';
-import { RestApi } from 'aws-cdk-lib/aws-apigateway';
+import { JobsApiConstruct } from '../../constructs/api/jobsConstruct';
 
 export class JobsApiStack extends Stack {
     constructor (parent: App, id: string, props: ApiStackProperties) {
@@ -28,144 +25,8 @@ export class JobsApiStack extends Stack {
             ...props,
         });
 
-        // Get common layer based on arn from SSM due to issues with cross stack references
-        const commonLambdaLayer = LayerVersion.fromLayerVersionArn(
-            this,
-            'mls-common-lambda-layer',
-            StringParameter.valueForStringParameter(this, props.mlspaceConfig.COMMON_LAYER_ARN_PARAM)
-        );
+        // eslint-disable-next-line @typescript-eslint/no-unused-vars
+        const jobsApiConstruct = new JobsApiConstruct(this, id + 'Resources', props);
 
-        const restApi = RestApi.fromRestApiAttributes(this, 'RestApi', {
-            restApiId: props.restApiId,
-            rootResourceId: props.rootResourceId,
-        });
-
-        const apis: MLSpacePythonLambdaFunction[] = [
-            {
-                name: 'list_training_jobs',
-                resource: 'hpo_job',
-                description: 'List HPO training jobs',
-                path: 'job/hpo/{jobName}/training-jobs',
-                method: 'GET',
-            },
-            {
-                name: 'describe',
-                resource: 'training_job',
-                description: 'Returns the description of a given training job',
-                path: 'job/training/{jobName}',
-                method: 'GET',
-            },
-            {
-                name: 'create',
-                resource: 'training_job',
-                description: 'Creates a new training job',
-                path: 'job/training',
-                method: 'POST',
-                environment: {
-                    ROLE_ARN: props.notebookInstanceRole.roleArn,
-                    BUCKET: props.configBucketName,
-                    S3_KEY: props.notebookParamFileKey,
-                    ENVIRONMENT: props.deploymentEnvironmentName,
-                },
-            },
-            {
-                name: 'create',
-                resource: 'transform_job',
-                description: 'Creates a transform job',
-                path: 'job/transform',
-                method: 'POST',
-                environment: {
-                    BUCKET: props.configBucketName,
-                    S3_KEY: props.notebookParamFileKey,
-                    ENVIRONMENT: props.deploymentEnvironmentName,
-                    DATA_BUCKET: props.dataBucketName,
-                },
-            },
-            {
-                name: 'stop',
-                resource: 'transform_job',
-                description: 'Stop transform job',
-                path: 'job/transform/{jobName}/stop',
-                method: 'POST',
-            },
-            {
-                name: 'describe',
-                resource: 'transform_job',
-                description: 'Describes a transform job',
-                path: 'job/transform/{jobName}',
-                method: 'GET',
-            },
-            {
-                name: 'create',
-                resource: 'hpo_job',
-                description: 'Creates a HPO job',
-                path: 'job/hpo',
-                method: 'POST',
-                environment: {
-                    BUCKET: props.configBucketName,
-                    S3_KEY: props.notebookParamFileKey,
-                    ENVIRONMENT: props.deploymentEnvironmentName,
-                },
-            },
-            {
-                name: 'stop',
-                resource: 'hpo_job',
-                description: 'Stop HPO job',
-                path: 'job/hpo/{jobName}/stop',
-                method: 'POST',
-            },
-            {
-                name: 'describe',
-                resource: 'hpo_job',
-                description: 'Describes a HPO job',
-                path: 'job/hpo/{jobName}',
-                method: 'GET',
-            },
-            {
-                id: 'job-get-logs',
-                name: 'get',
-                resource: 'logs',
-                description: 'Returns the log events for the specified job',
-                path: 'job/{jobType}/{jobName}/logs',
-                method: 'GET',
-            },
-            {
-                name: 'describe',
-                resource: 'labeling_job',
-                description: 'Describes a Ground Truth labeling job',
-                path: 'job/labeling/{jobName}',
-                method: 'GET',
-            },
-            {
-                name: 'create',
-                resource: 'labeling_job',
-                description: 'Creates a Labeling job',
-                path: 'job/labeling',
-                method: 'POST',
-                environment: {
-                    BUCKET: props.configBucketName,
-                    S3_KEY: props.notebookParamFileKey,
-                    ENVIRONMENT: props.deploymentEnvironmentName,
-                },
-            },
-        ];
-
-        apis.forEach((f) => {
-            registerAPIEndpoint(
-                this,
-                restApi,
-                props.authorizer,
-                props.applicationRole,
-                props.applicationRole.roleName,
-                props.notebookInstanceRole.roleName,
-                props.lambdaSourcePath,
-                [commonLambdaLayer],
-                f,
-                props.mlSpaceVPC,
-                props.securityGroups,
-                props.mlspaceConfig,
-                props.permissionsBoundaryArn
-            );
-        });
     }
 }
